Silence the eleven warnings the library still built with
All eleven are the compiler reading the code correctly; none is a defect,
and renders of canyon, bigben and Africa are byte-identical either way.

Two `-Wdouble-promotion` sites in the interpolator and two more in the
tone generator's bus probe were float operands meeting a double, which is
what the accumulators want -- the promotion is now spelled out rather than
left implicit, so the flag stops reporting a conversion nobody has to
think about and would still report one nobody intended.

Six `-Wshadow` sites were two different quantities wearing one name.  The
per-voice sends multiplied by their key scales were `reverb_level` and
siblings, which are the system units' own master levels one scope out;
they are `effective_*_send` now.  The drum-setup plane locals were bare
`reverb`/`chorus`/`delay`, which are the effect units themselves; `kit_`
says which is meant.

`to_pcm16` was an unused static in every build configured without libFLAC,
the FLAC encoder being its only caller.  It moves inside that `#ifdef`,
next to the anonymous namespace already there.

+++ src/dsp/interpolator.cpp
@@ -138,7 +138,7 @@ float SincInterpolator::sample(std::span<const float> buffer,
             continue;
         }
         const std::int64_t at = std::clamp(base + m, std::int64_t{0}, last);
-        sum += buffer[static_cast<std::size_t>(at)] * w;
+        sum += static_cast<double>(buffer[static_cast<std::size_t>(at)]) * w;
         density += w;
     }
     // Normalised by the weights actually summed, so every fractional position has unity DC gain
@@ -161,7 +161,7 @@ float SincInterpolator::sample_ring(std::span<const float> ring,
         if (w == 0.0) {
             continue;
         }
-        sum += ring[static_cast<std::size_t>((index + m) & mask)] * w;
+        sum += static_cast<double>(ring[static_cast<std::size_t>((index + m) & mask)]) * w;
         density += w;
     }
     return density > 0.0 ? static_cast<float>(sum / density) : 0.0F;
